-build-container.yml,+build-and-test-container.yml: also run tests
And make sure that we run the workflow all the time, but that in PRs we do
not build the container unnecessarily.

This is not the most optimized way to execute things. Since we first, build,
push, and then run the tests. Meaning that if the tests execute, the container
was already pushed. But it is nonetheless the most straight-forward way
to execute the whole process. Dealing with the container is very, very
annoying due to its size and complexity, and github actions are extremely
rigid and have many limitation with respect of when to run the workflows,
and dealing with containers in general. So as a first approach on which to
further iterate, this should certainly work.

Codeveloped-by: Pablo Correa Gomez <pablo.correa.gomez@combine.se>